style: fix clang-tidy warnings across source/*.cpp
Resolve all clang-tidy warnings reported by run-clang-tidy in source/*.cpp
(header warnings are out of scope per project policy).

- readability-static-accessed-through-instance: call CostHelper static
  methods via the class (CostHelper::stage_ecost/block_ecost/cost_factor)
  instead of through SystemContext instances. Drop now-unused parameters
  from file-local helpers in cost_helper.cpp and reserve_provision_lp.cpp.
- misc-const-correctness: add const to local variables in lp_fingerprint,
  sddp_cut_io, sddp_state_io.
- modernize-use-scoped-lock: replace lock_guard with scoped_lock in
  sddp_aperture.cpp.
- hicpp-signed-bitwise: use unsigned shift literal in linear_problem.cpp.
- readability-redundant-casting: drop redundant casts in sddp_state_io.cpp.
- performance-faster-string-find: prefer char over single-char string in
  solver_status.cpp.
- Add targeted NOLINT suppressions for unavoidable pointer-arithmetic,
  empty-catch, and recursion warnings in linear_interface, memory_compress,
  and memory_monitor.

+++ source/sddp_state_io.cpp
@@ -81,9 +81,7 @@ auto save_state_csv(const PlanningLP& planning_lp,
             continue;
           }
           const auto phys_val = col_sol[ci];
-          const auto rc = c < static_cast<size_t>(col_rc.size())
-              ? static_cast<double>(col_rc[ci])
-              : 0.0;
+          const auto rc = c < col_rc.size() ? col_rc[ci] : 0.0;
 
           ofs << names[ci] << "," << phase_uid << "," << scene_uid << ","
               << phys_val << "," << rc << "\n";
